+++ src/plot/jim.jl
@@ -10,7 +10,7 @@ using Plots: heatmap, plot!, annotate!
 using LaTeXStrings
 using MosaicViews: mosaicview
 using FFTViews: FFTView
-using Test: @test
+
 
 # global default key/values
 jim_def = Dict([
@@ -254,24 +254,6 @@ function jim(test::Symbol)
 	if haskey(jim_def, test)
 		return jim_def[test]
 	end
-	test != :test && throw("symbol $test")
-	jim()
-	jim(:keys)
-	jim(:clim)
-	@test typeof(jim(:defs)) <: Dict
 
-	jim(ones(4,3), title="test2", xlabel=L"x")
-	jim(rand(4,3,5), title=L"test3 x^2_i")
-	jim(1:4, 5:9, zeros(4,5), title="test3", ylabel=L"y")
-	jim(zeros(4,5), x=1:4, y=5:9, title="test3")
-	jim(rand(6,4), fft0=true)
-	jim(x=1:4, y=5:9, rand(4,5), title="test4")
-	jim(x=-9:9, y=9:-1:-9, (-9:9) * (abs.((9:-1:-9) .- 5) .< 3)', title="rev")
-	jim(rand(4,5), color=:hsv)
-	jim(ones(3,3)) # uniform
-	jim(:abswarn, false)
-	jim(complex(rand(4,3)))
-	jim(complex(rand(4,3)), "complex")
-	jim(:abswarn, true)
-	true
+	throw("symbol $test")
 end
