--- src/plot/z-test.jl
@@ -1,5 +1,0 @@
-# plot/z-test.jl
-
-using Test
-
-@test jim(:test)
