+++ src/regularize/Aodwt.jl
@@ -5,11 +5,8 @@ Aodwt.jl
 
 export Aodwt
 
-using Plots
 using LinearMapsAA: LinearMapAA, LinearMapAM, LinearMapAO
 using Wavelets: dwt!, idwt!, wavelet, WT
-using Test: @test
-#using MIRT: jim
 
 
 """
@@ -80,49 +77,4 @@ function Aodwt(
 	end
 
 	return A, scales, mfun
-end
-
-
-function Aodwt()
-	@doc Aodwt
-end
-
-
-"""
-    Aodwt_show( ; dims::Dims=(64,32), level::Int=3)
-show scales
-"""
-function Aodwt_show( ; dims::Dims = (64, 32), level::Int=3)
-	W, scales, mfun = Aodwt(dims, level=level)
-	jim(scales)
-end
-
-
-"""
-    Aodwt(:test)
-self test
-
-`Aodwt(:show)`
-visualize
-"""
-function Aodwt(test::Symbol)
-	if test === :show
-		return Aodwt_show()
-	end
-
-	test != :test && throw(ArgumentError("test $test"))
-	Aodwt()
-
-	for op in (true, false)
-		W,_,_ = Aodwt((8,16) ; level=2, operator=op)
-		@test Matrix(W)' == Matrix(W') # check adjoint
-		op && isinteractive() && (@show W.wt)
-		op && isinteractive() && (@show propertynames(W.wt))
-		@test W.level == 2
-		@test W.wt.name == "haar"
-		@test W isa (op ? LinearMapAO : LinearMapAM)
-	end
-
-	Aodwt(:show)
-	true
 end
