--- src/system/z-test.jl
@@ -1,5 +1,0 @@
-# system/z-test.jl
-
-using Test: @test
-
-@test Afft(:test)
