+++ src/utility/max_percent_diff.jl
@@ -5,8 +5,6 @@ Copyright 2000-9-16, Jeff Fessler, University of Michigan
 
 export max_percent_diff
 
-using Test: @test
-
 
 """
     d = max_percent_diff(s1, s2, [options])
@@ -22,7 +20,7 @@ options
 function max_percent_diff(s1, s2 ;
         maxboth::Bool = false, normalize::Bool = false)
 
-    #check to make sure that we have comparable signals
+    # check to make sure that we have comparable signals
     size(s1) != size(s2) &&
         throw(DomainError("Dimension mismatch $(size(s1)) $(size(s2))"))
 
@@ -47,25 +45,3 @@ function max_percent_diff(s1, s2 ;
 
     return 100 * d
 end
-
-
-function max_percent_diff_test()
-    @test max_percent_diff([200], [202]) ≈ 100/101
-    @test max_percent_diff([0 200], [0 202]) ≈ 100/101
-    @test max_percent_diff([0 200], [0 198], maxboth=true) ≈ 1
-    #≈ does not work with 0 values (it's proportional to the second arg)
-    @test max_percent_diff([0 100], [0 200], normalize=true) < .01
-    true
-end
-
-
-"""
-    max_percent_diff(:test)
-self test
-"""
-function max_percent_diff(test::Symbol)
-    test != :test && throw(DomainError("Invalid signal"));
-    return max_percent_diff_test()
-end
-
-# max_percent_diff(:test)
