+++ test/plot/z-test.jl
@@ -1,0 +1,14 @@
+# plot/z-test.jl
+
+using Test: @testset
+
+list = [
+"jim.jl"
+#"setup.jl"
+]
+
+for file in list
+	@testset "$file" begin
+		include(file)
+	end
+end
